Fixed unexposed ushort bit shift.
• Fixed a ushort right-shift which had a regular integer as the right-hand operand.

+++ Assets/Nessie/Udon/NUSaveState/Scripts/U#/NUSaveState.cs
@@ -532,7 +532,7 @@ namespace Nessie.Udon.SaveState
                 ushort bytes = _ReadParameter(boneIndex);
                 output[byteIndex++ + avatarByteOffset] = (byte)(bytes & 0xFF);
                 if (byteIndex < avatarByteCount)
-                    output[byteIndex++ + avatarByteOffset] = (byte)(bytes >> 8);
+                    output[byteIndex++ + avatarByteOffset] = (byte)(bytes >> (ushort)8);
             }
 
             return output;
